examples/fspbt/requirements.txt
numpy
opencv-python
Pillow
PyYAML
scikit-image
scipy
tensorboard
torch
torchvision